(Refactor) backend
Migrate database creation code from db.js to test/index.js. Running 'npm run load' will now check for existence of database, create, and populate with example data.  Removed console.log statements from patch routes
